pyproject.toml
[project]
name = "web-scraper-voice-agent"
version = "0.1.0"
description = "A web scraper with voice agent capabilities"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "openai>=1.51.0",
    "langchain_openai>=0.1.103",
    "langchain>=0.3.2,<0.4.0",
    "transformers>=4.40.0",
    "torch>=2.1.0",
    "accelerate>=0.23.0",
    "sentence-transformers>=2.2.2",
    "dependency-injector>=4.48.3",
    "pdfplumber>=0.11.9",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.2",
    "python-multipart>=0.0.21",
    "langgraph>=0.2.0"
]
requires-python = ">=3.10,<3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [] 